refactor(indexer-alt): ev indices explicit NOT NULL on key columns
## Description

Add explicit `NOT NULL`s to the key columns inside the event filter
tables. This is not a behavioural change as these columns are already
`NOT NULL` because they are part of the primary key. As per an earlier
PR comment, relying on the implicit `NOT NULL` behaviour was confusing
when reading migrations (and other migration scripts use the explicit
pattern).

## Test plan

CI